requirements.txt
faiss-cpu
langchain
langchain-community
langchain-core
PyPDF2
google-cloud-aiplatform
python-dotenv
tiktoken
